Fix TMinor::ReturnFalseNationStateCapabilityFlag90 body and slot param type
The TMinor override (0x4e45f0) carried the wrong body — a 0xd..0x10 range check
(copied from the unrelated IsSpecialNationInteractionResource predicate) — scoring
9.5%. The real function tests whether the argument equals any of the four saved
diplomacy nation slots at diplomacySaveFields134[0..3], using the init-to-0 /
set-to-1 single-return form. The vtable slot argument is a short (word compares),
not the int the base decl used; narrowing the slot param across TCountry (base,
return-false, stays 100%) and TMinor (override) in lockstep produces the word load.

TMinor::ReturnFalseNationStateCapabilityFlag90 9.5%->100%; TCountry base unchanged
at 100%. aligned 2392->2393 (+1), no regressions.

+++ include/game/TMinor.h
@@ -53,7 +53,7 @@ public:
   char TryDispatchNationActionViaUiContextOrFallback(int arg1, int arg2, int arg3,
                                                      int arg4) override;
   void QueueDiplomacyProposalCodeForTargetNation(short proposalCode, short targetNationId) override;
-  char ReturnFalseNationStateCapabilityFlag90(int arg) override;
+  char ReturnFalseNationStateCapabilityFlag90(short arg) override;
   void NotifyActionSlot94(int sourceNation, int actionCode) override;
 
   // slot 0x2a (+0xa8), TMinor's first new virtual — vtable 0x653c90+0xa8 -> 0x4e46a0,
